MedicalPromptGeneration/requirements.txt
Pillow
tqdm
numpy
transformers
torch>=1.11.0
requests